requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
requests>=2.28.0
python-dotenv>=1.0.0
